refactor(dht): share self-address publish policy between driver and DHT self-entry
Extracts the typed self-address builder out of the reachability driver
into a new `self_address` module so `DhtNetworkManager::local_dht_node()`
and the driver's `PublishAddressSet` fan-out apply identical invariants:
relay first when present, at most one best non-relay address per IP
family, Direct only when the passive classifier has proven the external
socket, wildcard and zero-port addresses dropped. Adds a
`relay_external_address()` accessor on `TransportHandle` so the DHT
self-entry can include the relay tier without duplicating storage.
